requirements.txt
# Dependencias para Quadra - agrupadas y comentadas

# -----------------------------
# Framework y extensiones de Flask
# (core, ORM, migraciones, autenticación y formularios)
# -----------------------------
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.5.2

# -----------------------------
# Base de datos / migraciones / driver
# -----------------------------
alembic==1.16.4
psycopg2-binary==2.9.10
SQLAlchemy==2.0.42
greenlet==3.2.4

# -----------------------------
# Seguridad y autenticación
# -----------------------------
bcrypt==4.0.1
PyJWT==2.8.0
itsdangerous==2.2.0

# -----------------------------
# Formularios y validación
# -----------------------------
WTForms==3.0.1
email-validator==2.0.0

# -----------------------------
# Manejo de imágenes y archivos
# -----------------------------
Pillow==10.0.0

# -----------------------------
# HTTP, requests y utilidades de red
# -----------------------------
requests==2.31.0
urllib3==2.5.0
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10

# -----------------------------
# CLI, entorno y utilidades de desarrollo
# -----------------------------
click==8.2.1
python-dotenv==1.0.0
colorama==0.4.6
packaging==25.0
typing_extensions==4.14.1
six==1.17.0

# -----------------------------
# Plantillas y renderizado
# -----------------------------
Jinja2==3.1.6
MarkupSafe==3.0.2
Mako==1.3.10

# -----------------------------
# Servidor de producción
# -----------------------------
gunicorn==21.2.0

# -----------------------------
# Auxiliares
# -----------------------------
dnspython==2.7.0
blinker==1.9.0

# Nota: pip freeze lista dependencias transitivas también. Mantén
# este archivo actualizado desde el virtualenv del proyecto.
